+++ great_expectations/checkpoint/types/checkpoint_result.py
@@ -5,7 +5,6 @@ from typing import Dict, List, Optional, Union
 from great_expectations.core.expectation_validation_result import (
     ExpectationSuiteValidationResult,
 )
-from great_expectations.core.id_dict import BatchKwargs
 from great_expectations.core.run_identifier import RunIdentifier, RunIdentifierSchema
 from great_expectations.core.util import convert_to_json_serializable
 from great_expectations.data_context.types.base import (
