+++ great_expectations/expectations/metrics/column_aggregate_metrics/column_histogram.py
@@ -4,7 +4,10 @@ from typing import Any, Dict, Tuple
 
 import numpy as np
 
-from great_expectations.core.util import convert_to_json_serializable
+from great_expectations.core.util import (
+    convert_to_json_serializable,
+    get_sql_dialect_floating_point_infinity_value,
+)
 from great_expectations.execution_engine import (
     PandasExecutionEngine,
     SparkDFExecutionEngine,
@@ -16,9 +19,6 @@ from great_expectations.expectations.metrics.column_aggregate_metric import (
 )
 from great_expectations.expectations.metrics.import_manager import Bucketizer, F, sa
 from great_expectations.expectations.metrics.metric_provider import metric_value
-from great_expectations.expectations.metrics.util import (
-    get_sql_dialect_floating_point_infinity_value,
-)
 
 logger = logging.getLogger(__name__)
