+++ tests/actions/test_validation_operators_in_data_context.py
@@ -2,11 +2,10 @@ import json
 
 import pytest
 
-from great_expectations.core.expectation_suite import ExpectationSuiteSchema
 from great_expectations.data_context import BaseDataContext
 from great_expectations.data_context.util import file_relative_path
 from great_expectations.exceptions import DataContextError
-from tests.test_utils import expectationSuiteSchema
+from great_expectations.self_check.util import expectationSuiteSchema
 
 
 @pytest.fixture()
