+++ tests/conftest.py
@@ -4,7 +4,6 @@ import locale
 import os
 import random
 import shutil
-import threading
 from types import ModuleType
 from typing import Dict, List, Optional
 
@@ -35,12 +34,14 @@ from great_expectations.dataset.pandas_dataset import PandasDataset
 from great_expectations.datasource import SqlAlchemyDatasource
 from great_expectations.datasource.new_datasource import Datasource
 from great_expectations.execution_engine import SqlAlchemyExecutionEngine
-from great_expectations.util import import_library_module
-from tests.test_utils import (
-    create_files_in_directory,
+from great_expectations.self_check.util import (
+    LockingConnectionCheck,
+    expectationSuiteSchema,
     expectationSuiteValidationResultSchema,
     get_dataset,
 )
+from great_expectations.util import import_library_module
+from tests.test_utils import create_files_in_directory
 
 yaml = YAML()
 ###
@@ -50,27 +51,6 @@ yaml = YAML()
 ###
 
 locale.setlocale(locale.LC_ALL, "en_US.UTF-8")
-
-
-class LockingConnectionCheck:
-    def __init__(self, sa, connection_string):
-        self.lock = threading.Lock()
-        self.sa = sa
-        self.connection_string = connection_string
-        self._is_valid = None
-
-    def is_valid(self):
-        with self.lock:
-            if self._is_valid is None:
-                try:
-                    engine = self.sa.create_engine(self.connection_string)
-                    conn = engine.connect()
-                    conn.close()
-                    self._is_valid = True
-                except (ImportError, self.sa.exc.SQLAlchemyError) as e:
-                    print(f"{str(e)}")
-                    self._is_valid = False
-            return self._is_valid
 
 
 def pytest_configure(config):
@@ -3900,7 +3880,7 @@ def sqlite_view_engine(test_backends):
 
             sqlite_engine = sa.create_engine("sqlite://")
             df = pd.DataFrame({"a": [1, 2, 3, 4, 5]})
-            df.to_sql("test_table", con=sqlite_engine)
+            df.to_sql(name="test_table", con=sqlite_engine, index=True)
             sqlite_engine.execute(
                 "CREATE TEMP VIEW test_temp_view AS SELECT * FROM test_table where a < 4;"
             )
@@ -3981,8 +3961,8 @@ def test_db_connection_string(tmp_path_factory, test_backends):
         basepath = str(tmp_path_factory.mktemp("db_context"))
         path = os.path.join(basepath, "test.db")
         engine = sa.create_engine("sqlite:///" + str(path))
-        df1.to_sql("table_1", con=engine, index=True)
-        df2.to_sql("table_2", con=engine, index=True, schema="main")
+        df1.to_sql(name="table_1", con=engine, index=True)
+        df2.to_sql(name="table_2", con=engine, index=True, schema="main")
 
         # Return a connection string to this newly-created db
         return "sqlite:///" + str(path)
@@ -4081,7 +4061,7 @@ SELECT EXISTS (
 """
     ).fetchall()
     if table_check_results != [(True,)]:
-        test_df.to_sql("test_df", con=engine, index=True, schema="connection_test")
+        test_df.to_sql(name="test_df", con=engine, index=True, schema="connection_test")
 
     # Return a connection string to this newly-created db
     return engine
