+++ tests/data_asset/test_data_asset_util.py
@@ -9,8 +9,7 @@ import numpy as np
 import pytest
 
 import great_expectations as ge
-from great_expectations.core.expectation_suite import ExpectationSuiteSchema
-from tests.test_utils import expectationSuiteSchema
+from great_expectations.self_check.util import expectationSuiteSchema
 
 
 def test_recursively_convert_to_json_serializable():
