+++ tests/dataset/test_pandas_dataset.py
@@ -7,7 +7,10 @@ import pytest
 import great_expectations as ge
 from great_expectations.core.expectation_configuration import ExpectationConfiguration
 from great_expectations.profile import ColumnsExistProfiler
-from tests.test_utils import expectationSuiteSchema, expectationValidationResultSchema
+from great_expectations.self_check.util import (
+    expectationSuiteSchema,
+    expectationValidationResultSchema,
+)
 
 
 def test_expect_column_values_to_be_dateutil_parseable():
