+++ tests/execution_engine/test_execution_engine.py
@@ -2,7 +2,7 @@ import pandas as pd
 import pytest
 
 from great_expectations.exceptions import GreatExpectationsError
-from great_expectations.execution_engine import ExecutionEngine, PandasExecutionEngine
+from great_expectations.execution_engine import PandasExecutionEngine
 from great_expectations.validator.validation_graph import MetricConfiguration
