+++ tests/execution_engine/test_sqlalchemy_execution_engine.py
@@ -4,7 +4,7 @@ import os
 import pandas as pd
 import pytest
 
-from great_expectations.core.batch import Batch, BatchSpec
+from great_expectations.core.batch import BatchSpec
 from great_expectations.data_context.util import file_relative_path
 from great_expectations.exceptions import GreatExpectationsError
 from great_expectations.exceptions.exceptions import InvalidConfigError
@@ -13,16 +13,10 @@ from great_expectations.execution_engine.execution_engine import MetricDomainTyp
 from great_expectations.execution_engine.sqlalchemy_execution_engine import (
     SqlAlchemyExecutionEngine,
 )
-from great_expectations.expectations.metrics import (
-    ColumnMean,
-    ColumnStandardDeviation,
-    ColumnValuesInSet,
-    ColumnValuesZScore,
-)
-from great_expectations.validator.validation_graph import MetricConfiguration
 
 # Function to test for spark dataframe equality
-from tests.test_utils import _build_sa_engine
+from great_expectations.self_check.util import build_sa_engine
+from great_expectations.validator.validation_graph import MetricConfiguration
 
 
 def test_instantiation_via_connection_string(sa, test_db_connection_string):
@@ -106,7 +100,7 @@ def test_instantiation_error_states(sa, test_db_connection_string):
 def test_sa_batch_aggregate_metrics(caplog, sa):
     import datetime
 
-    engine = _build_sa_engine(
+    engine = build_sa_engine(
         pd.DataFrame({"a": [1, 2, 1, 2, 3, 3], "b": [4, 4, 4, 4, 4, 4]}), sa
     )
 
@@ -195,7 +189,7 @@ def test_sa_batch_aggregate_metrics(caplog, sa):
 
 # Ensuring functionality of compute_domain when no domain kwargs are given
 def test_get_compute_domain_with_no_domain_kwargs(sa):
-    engine = _build_sa_engine(
+    engine = build_sa_engine(
         pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None]}), sa
     )
 
@@ -217,7 +211,7 @@ def test_get_compute_domain_with_no_domain_kwargs(sa):
 
 # Testing for only untested use case - column_pair
 def test_get_compute_domain_with_column_pair(sa):
-    engine = _build_sa_engine(
+    engine = build_sa_engine(
         pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None]}), sa
     )
 
@@ -244,7 +238,7 @@ def test_get_compute_domain_with_column_pair(sa):
     }, "Accessor kwargs have been modified"
 
     # Building new engine so that values still found
-    engine = _build_sa_engine(
+    engine = build_sa_engine(
         pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None]}), sa
     )
     data2, compute_kwargs, accessor_kwargs = engine.get_compute_domain(
@@ -270,7 +264,7 @@ def test_get_compute_domain_with_column_pair(sa):
 
 # Testing for only untested use case - multicolumn
 def test_get_compute_domain_with_multicolumn(sa):
-    engine = _build_sa_engine(
+    engine = build_sa_engine(
         pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None], "c": [1, 2, 3, None]}),
         sa,
     )
@@ -316,7 +310,7 @@ def test_get_compute_domain_with_multicolumn(sa):
 
 # Testing whether compute domain is properly calculated, but this time obtaining a column
 def test_get_compute_domain_with_column_domain(sa):
-    engine = _build_sa_engine(
+    engine = build_sa_engine(
         pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None]}), sa
     )
 
@@ -337,7 +331,7 @@ def test_get_compute_domain_with_column_domain(sa):
     assert accessor_kwargs == {"column": "a"}, "Accessor kwargs have been modified"
 
     # Testing for identity
-    engine = _build_sa_engine(
+    engine = build_sa_engine(
         pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None]}), sa
     )
 
@@ -360,7 +354,7 @@ def test_get_compute_domain_with_column_domain(sa):
 
 # What happens when we filter such that no value meets the condition?
 def test_get_compute_domain_with_unmeetable_row_condition(sa):
-    engine = _build_sa_engine(
+    engine = build_sa_engine(
         pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None]}), sa
     )
 
@@ -391,7 +385,7 @@ def test_get_compute_domain_with_unmeetable_row_condition(sa):
 
 # Testing to ensure that great expectation experimental parser also works in terms of defining a compute domain
 def test_get_compute_domain_with_ge_experimental_condition_parser(sa):
-    engine = _build_sa_engine(
+    engine = build_sa_engine(
         pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None]}), sa
     )
 
@@ -434,7 +428,7 @@ def test_get_compute_domain_with_ge_experimental_condition_parser(sa):
 
     # Ensuring data has been properly queried
     # Seeing if raw data is the same as the data after condition has been applied - checking post computation data
-    engine = _build_sa_engine(
+    engine = build_sa_engine(
         pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None]}), sa
     )
     raw_data = engine.engine.execute(
@@ -452,7 +446,7 @@ def test_get_compute_domain_with_ge_experimental_condition_parser(sa):
 
 
 def test_get_compute_domain_with_nonexistent_condition_parser(sa):
-    engine = _build_sa_engine(
+    engine = build_sa_engine(
         pd.DataFrame({"a": [1, 2, 3, 4], "b": [2, 3, 4, None]}), sa
     )
 
@@ -469,7 +463,7 @@ def test_get_compute_domain_with_nonexistent_condition_parser(sa):
 
 # Ensuring that we can properly inform user when metric doesn't exist - should get a metric provider error
 def test_resolve_metric_bundle_with_nonexistent_metric(sa):
-    engine = _build_sa_engine(
+    engine = build_sa_engine(
         pd.DataFrame({"a": [1, 2, 1, 2, 3, 3], "b": [4, 4, 4, 4, 4, 4]}), sa
     )
