+++ tests/render/test_render.py
@@ -4,7 +4,6 @@ from collections import OrderedDict
 import pytest
 
 import great_expectations as ge
-from great_expectations.core.expectation_suite import ExpectationSuiteSchema
 from great_expectations.data_context.util import file_relative_path
 from great_expectations.profile.basic_dataset_profiler import BasicDatasetProfiler
 from great_expectations.render.renderer import (
@@ -19,7 +18,7 @@ from great_expectations.render.renderer.content_block import (
     ValidationResultsTableContentBlockRenderer,
 )
 from great_expectations.render.view import DefaultJinjaPageView
-from tests.test_utils import (
+from great_expectations.self_check.util import (
     expectationSuiteSchema,
     expectationSuiteValidationResultSchema,
 )
