+++ tests/test_definitions/test_expectations.py
@@ -11,31 +11,17 @@ import pandas as pd
 import pytest
 
 from great_expectations.dataset import PandasDataset, SparkDFDataset, SqlAlchemyDataset
-
-from ..conftest import build_test_backends_list
-from ..test_utils import (
+from great_expectations.self_check.util import (
     candidate_test_is_on_temporary_notimplemented_list,
     evaluate_json_test,
     get_dataset,
+    mssqlDialect,
+    mysqlDialect,
+    postgresqlDialect,
+    sqliteDialect,
 )
 
-try:
-    from sqlalchemy.dialects.mssql import dialect as mssqlDialect
-except (ImportError, KeyError):
-    mssqlDialect = None
-try:
-    from sqlalchemy.dialects.mysql import dialect as mysqlDialect
-except (ImportError, KeyError):
-    mysqlDialect = None
-try:
-    from sqlalchemy.dialects.postgresql import dialect as postgresqlDialect
-except (ImportError, KeyError):
-    postgresqlDialect = None
-try:
-    from sqlalchemy.dialects.sqlite import dialect as sqliteDialect
-except (ImportError, KeyError):
-    sqliteDialect = None
-
+from ..conftest import build_test_backends_list
 
 logger = logging.getLogger(__name__)
 tmp_dir = str(tempfile.mkdtemp())
@@ -59,8 +45,6 @@ def pytest_generate_tests(metafunc):
             dir_path + "/" + expectation_category + "/*.json"
         )
         for c in build_test_backends_list(metafunc):
-            if c != "SparkDFDataset":
-                continue
             for filename in test_configuration_files:
                 file = open(filename)
                 # Use OrderedDict so that python2 will use the correct order of columns in all cases
