+++ tests/test_definitions/test_expectations_cfe.py
@@ -12,7 +12,11 @@ from great_expectations.execution_engine.sparkdf_batch_data import SparkDFBatchD
 from great_expectations.execution_engine.sqlalchemy_batch_data import (
     SqlAlchemyBatchData,
 )
-from great_expectations.expectations.util import get_test_validator_with_data
+from great_expectations.self_check.util import (
+    candidate_test_is_on_temporary_notimplemented_list_cfe,
+    evaluate_json_test_cfe,
+    get_test_validator_with_data,
+)
 from tests.conftest import build_test_backends_list_cfe
 from tests.test_definitions.test_expectations import mssqlDialect as mssqlDialect
 from tests.test_definitions.test_expectations import mysqlDialect as mysqlDialect
@@ -21,10 +25,6 @@ from tests.test_definitions.test_expectations import (
 )
 from tests.test_definitions.test_expectations import sqliteDialect as sqliteDialect
 from tests.test_definitions.test_expectations import tmp_dir
-from tests.test_utils import evaluate_json_test_cfe
-from tests.test_utils_modular import (
-    candidate_test_is_on_temporary_notimplemented_list_cfe,
-)
 
 
 def pytest_generate_tests(metafunc):
@@ -46,8 +46,6 @@ def pytest_generate_tests(metafunc):
             dir_path + "/" + expectation_category + "/*.json"
         )
         for c in backends:
-            if c != "spark":
-                continue
             for filename in test_configuration_files:
                 file = open(filename)
                 test_configuration = json.load(file)
